Splitting out test to reduce memory use
